requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.8.0
Django>=4.2.0
exceptiongroup==1.2.2
factory_boy>=3.3.0
Faker>=19.3.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.5.0
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-django>=4.5.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
tzdata==2025.2
selenium>=4.15.2
pytest-factoryboy>=2.5.1
